fix(ui): qualify layout params and layout directions for the card hosts
The accordion and the action row are hosted by MaterialCardView, so unqualified LayoutParams resolved to the card's own nested type instead of the LinearLayout child params, and the LinearLayout direction constants are no longer in scope: both now name the type they mean. The panel host reads the activity from the view rather than from the InlineHost interface, and the card radius is set as the float the framework expects.

+++ app/src/main/java/sumicya/qself/ui/SettingsVisuals.kt
@@ -189,7 +189,7 @@ object SettingsVisuals {
         container: Int = p.surfaceLow,
         clickable: Boolean = false,
     ): MaterialCardView = MaterialCardView(context).apply {
-        this.radius = dp(context, radius)
+        this.radius = dp(context, radius).toFloat()
         cardElevation = 0f
         strokeWidth = 0
         setCardBackgroundColor(container)
